<fix> itemCategory slice 이동
