+++ src/redux/store.js
@@ -3,7 +3,6 @@ import thunk from "redux-thunk";
 import marketPost from "./modules/market/postSlice";
 import user from "./modules/user/userSlice";
 import comment from "./modules/market/commentSlice";
-import category from "./modules/filter/categorySlice";
 
 const middlewares = [thunk];
 
@@ -12,7 +11,6 @@ const rootReducer = combineReducers({
   marketPost,
   user,
   comment,
-  category,
 });
 
 // 스토어 연결
